feat(auth): 重构登录逻辑并添加用户权限管理
- 新增 LoginUtil 类用于存储登录相关的常量
- 重构 TokenService 类,优化登录流程并添加用户权限相关方法
- 新增多个 DTO 类用于承载用户登录和权限信息
- 添加 RequestUtil 工具类用于获取请求相关的信息
- 更新相关的 mapper 和 service 接口

+++ xh-system/src/main/java/com/xh/system/infrastructure/mysql/po/SysRoleMenuPO.java
@@ -13,8 +13,9 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = true)
 public class SysRoleMenuPO extends BaseEntity {
     @Schema(title = "角色id")
-    private Integer sysRoleId;
+    private Long sysRoleId;
 
     @Schema(title = "菜单id")
-    private Integer sysMenuId;
+    private Long sysMenuId;
+    
 }
